+++ apps/frontend/app/lib/api.server.ts
@@ -6,8 +6,9 @@ import { withQuery } from "ufo";
 import { authCookie } from "~/lib/cookies.server";
 import { redirectToQueryParam } from "./generals";
 import { createToastHeaders } from "./toast.server";
+import { combineHeaders, getLogoutCookies } from "./utilities.server";
 
-export const API_URL = process.env.API_URL;
+export const API_URL = process.env.API_URL || "http://localhost:5000";
 
 export const gqlClient = new GraphQLClient(`${API_URL}/graphql`, {
 	headers: { Connection: "keep-alive" },
@@ -26,12 +27,16 @@ export const getAuthorizationHeader = async (request: Request) => {
 export const getIsAuthenticated = async (request: Request) => {
 	const cookie = await getAuthorizationCookie(request);
 	if (!cookie) return [false, null] as const;
-	const { userDetails } = await gqlClient.request(
-		UserDetailsDocument,
-		undefined,
-		await getAuthorizationHeader(request),
-	);
-	return [userDetails.__typename === "User", userDetails] as const;
+	try {
+		const { userDetails } = await gqlClient.request(
+			UserDetailsDocument,
+			undefined,
+			await getAuthorizationHeader(request),
+		);
+		return [userDetails.__typename === "User", userDetails] as const;
+	} catch {
+		return [false, null] as const;
+	}
 };
 
 export const redirectIfNotAuthenticated = async (request: Request) => {
@@ -44,9 +49,12 @@ export const redirectIfNotAuthenticated = async (request: Request) => {
 			}),
 			{
 				status: 302,
-				headers: await createToastHeaders({
-					message: "You must be logged in to view this page",
-				}),
+				headers: combineHeaders(
+					await createToastHeaders({
+						message: "You must be logged in to view this page",
+					}),
+					{ "Set-Cookie": await getLogoutCookies() },
+				),
 			},
 		);
 	}