+++ apps/frontend/app/routes/actions.tsx
@@ -13,9 +13,9 @@ import { namedAction } from "remix-utils/named-action";
 import { z } from "zod";
 import { zx } from "zodix";
 import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
-import { authCookie, colorSchemeCookie } from "~/lib/cookies.server";
+import { colorSchemeCookie } from "~/lib/cookies.server";
 import { createToastHeaders } from "~/lib/toast.server";
-import { processSubmission } from "~/lib/utilities.server";
+import { getLogoutCookies, processSubmission } from "~/lib/utilities.server";
 
 export const loader = async () => redirect($path("/"));
 
@@ -36,18 +36,19 @@ export const action = async ({ request }: ActionFunctionArgs) => {
 				request.headers.get("Cookie") || "",
 			);
 			const newColorScheme = currentColorScheme === "dark" ? "light" : "dark";
-			return redirect($path("/"), {
-				headers: {
-					"Set-Cookie": await colorSchemeCookie.serialize(newColorScheme),
+			return json(
+				{},
+				{
+					headers: {
+						"Set-Cookie": await colorSchemeCookie.serialize(newColorScheme),
+					},
 				},
-			});
+			);
 		},
 		logout: async () => {
 			return redirect($path("/auth/login"), {
 				headers: {
-					"Set-Cookie": await authCookie.serialize("", {
-						expires: new Date(0),
-					}),
+					"Set-Cookie": await getLogoutCookies(),
 				},
 			});
 		},